# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Athena Node API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/master-settings": {
      "get": {
        "tags": [
          "master-settings"
        ],
        "summary": "Get Master Settings",
        "operationId": "get_master_settings_api_v1_master_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterSettingResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "master-settings"
        ],
        "summary": "Update Master Settings",
        "operationId": "update_master_settings_api_v1_master_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MasterSettingInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterSettingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/master-settings/test": {
      "post": {
        "tags": [
          "master-settings"
        ],
        "summary": "Test Master Settings",
        "operationId": "test_master_settings_api_v1_master_settings_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MasterSettingInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterConnectionTestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/master-settings/registration": {
      "post": {
        "tags": [
          "master-settings"
        ],
        "summary": "Submit Registration Application",
        "operationId": "submit_registration_application_api_v1_master_settings_registration_post",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationApplicationResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/master-settings/registration/status": {
      "post": {
        "tags": [
          "master-settings"
        ],
        "summary": "Synchronize Registration Status",
        "operationId": "synchronize_registration_status_api_v1_master_settings_registration_status_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MasterConnectionTestResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "default": "success"
          }
        },
        "type": "object",
        "title": "MasterConnectionTestResponse"
      },
      "MasterSettingInput": {
        "properties": {
          "scheme": {
            "type": "string",
            "enum": [
              "http",
              "https"
            ],
            "title": "Scheme"
          },
          "host": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Host"
          },
          "port": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Port"
          },
          "token": {
            "type": "string",
            "title": "Token",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "scheme",
          "host",
          "port"
        ],
        "title": "MasterSettingInput"
      },
      "MasterSettingResponse": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "node_name": {
            "type": "string",
            "title": "Node Name"
          },
          "scheme": {
            "type": "string",
            "enum": [
              "http",
              "https"
            ],
            "title": "Scheme"
          },
          "host": {
            "type": "string",
            "title": "Host"
          },
          "port": {
            "type": "integer",
            "title": "Port"
          },
          "has_token": {
            "type": "boolean",
            "title": "Has Token"
          },
          "runtime_status": {
            "type": "string",
            "enum": [
              "unconfigured",
              "connecting",
              "online",
              "error",
              "stopped"
            ],
            "title": "Runtime Status"
          },
          "registration_status": {
            "type": "string",
            "enum": [
              "not_submitted",
              "pending",
              "approved"
            ],
            "title": "Registration Status"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "node_name",
          "scheme",
          "host",
          "port",
          "has_token",
          "runtime_status",
          "registration_status"
        ],
        "title": "MasterSettingResponse"
      },
...
      "RegistrationApplicationResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "pending",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "RegistrationApplicationResponse"
      },
      "RegistrationStatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "RegistrationStatusResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}